// src/gen/api/types/ListOutgoingFriendRequests.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { FriendRequest } from './FriendRequest.ts'

export type ListOutgoingFriendRequestsQueryParams = {
  /**
   * @type string | undefined
   */
  cursor?: string
  /**
   * @minLength 1
   * @maxLength 50
   * @default 20
   * @type integer | undefined
   */
  limit?: number
}

/**
 * @description Cursor-paginated outgoing pending requests
 */
export type ListOutgoingFriendRequests200 = {
  /**
   * @type array
   */
  data: FriendRequest[]
  /**
   * @type object
   */
  pagination: {
    /**
     * @type string,null
     */
    nextCursor: string | null
    /**
     * @type boolean
     */
    hasNext: boolean
    /**
     * @type number
     */
    limit: number
  }
}

/**
 * @description Validation failed
 */
export type ListOutgoingFriendRequests400 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Missing or invalid access token
 */
export type ListOutgoingFriendRequests401 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Internal server error
 */
export type ListOutgoingFriendRequests500 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

export type ListOutgoingFriendRequestsQueryResponse =
  ListOutgoingFriendRequests200

export type ListOutgoingFriendRequestsQuery = {
  Response: ListOutgoingFriendRequests200
  QueryParams: ListOutgoingFriendRequestsQueryParams
  Errors:
    | ListOutgoingFriendRequests400
    | ListOutgoingFriendRequests401
    | ListOutgoingFriendRequests500
}